batch 再售

+++ voyageone-components/voyageone-components-jumei/src/main/java/com/voyageone/components/jumei/Bean/HtDeal_UpdateDealPriceBatch_UpdateData.java
@@ -8,8 +8,8 @@ import java.math.BigDecimal;
 public class HtDeal_UpdateDealPriceBatch_UpdateData {
    String jumei_sku_no;//Number  聚美Sku_no;
     String  jumei_hash_id;//	String    聚美Deal唯一值;
-    Float market_price;// 可选	Float            市场价;    参数范围: 注:市场价和团购价不能同时为空,市场价必须大于等于团购价
-    Float  deal_price;// 可选	Float       团购价; 参数范围: 注:市场价和团购价不能同时为空,团购价至少大于15元
+    double market_price;// 可选	Float            市场价;    参数范围: 注:市场价和团购价不能同时为空,市场价必须大于等于团购价
+    double  deal_price;// 可选	Float       团购价; 参数范围: 注:市场价和团购价不能同时为空,团购价至少大于15元
     public String getJumei_sku_no() {
         return jumei_sku_no;
     }
@@ -22,16 +22,16 @@ public class HtDeal_UpdateDealPriceBatch_UpdateData {
     public void setJumei_hash_id(String jumei_hash_id) {
         this.jumei_hash_id = jumei_hash_id;
     }
-    public Float getMarket_price() {
+    public double getMarket_price() {
         return market_price;
     }
-    public void setMarket_price(Float market_price) {
+    public void setMarket_price(double market_price) {
         this.market_price = market_price;
     }
-    public Float getDeal_price() {
+    public double getDeal_price() {
         return deal_price;
     }
-    public void setDeal_price(Float deal_price) {
+    public void setDeal_price(double deal_price) {
         this.deal_price = deal_price;
     }
 }
